Finit la cascade Decimal dans les harnais, et fait regarder le gate (#443)
#443 demandait de terminer la migration Decimal dans les tests BDD et e2e. Sa
partie compilation est résorbée depuis longtemps : `kcargo check
--all-targets` rend 0 erreur et 0 warning, et la CI exécute `Unit Tests`,
`Integration Tests` et `BDD Tests` au vert.

Mais la cascade elle-même ne l'était pas, et personne ne pouvait le voir.

── Ce que le gate ne regardait pas ────────────────────────────────────────

`scripts/check-no-f64-money.sh` prend un chemin en argument depuis l'origine.
La CI ne l'appelait qu'avec son défaut, `backend/src`. Les harnais de test
étaient donc hors de sa vue, et manipulaient l'argent autrement que le
produit :

    bdd.rs:92   last_invoice_vat_amount: Option<f64>   ← une TVA
    bdd.rs:94   last_invoice_total: Option<f64>        ← un total de facture
    bdd_financial.rs:3737  amount2: f64                ← lié à une colonne NUMERIC

Le troisième est le plus net : un montant d'euros, entier dans le gabarit
Gherkin, qui traversait le flottant avant d'atteindre une colonne exacte.

**Un test qui traite l'argent en flottant ne prouve pas que le produit ne le
fait pas.** Il prouve autre chose.

── Les corrections ────────────────────────────────────────────────────────

Les trois passent en `Decimal`. Le paramètre de step devient `u64` — le
gabarit capture `(\d+) EUR`, donc un entier — puis `Decimal::from` avant le
`bind`. Pas d'aller-retour flottant, ce que l'ADR-0008 §A proscrit
explicitement.

Le gate est appelé une seconde fois sur `backend/tests`. Vérifié par témoin :
`f64` réintroduit sur la TVA, code de sortie 1 ; retiré, code 0.

C'est la sixième garde inopérante de la journée, et la variante la plus
discrète : celle-ci n'était pas mal écrite ni court-circuitée, elle regardait
simplement ailleurs. Le script était juste, son appel incomplet.

`kcargo check --all-targets` : 0 erreur, 0 warning. `fmt --all` appliqué.
YAML validé.

+++ backend/tests/bdd_financial.rs
@@ -3734,7 +3734,11 @@ async fn then_all_distributions_for_alice(world: &mut FinancialWorld) {
 }
 
 #[given(regex = r#"^charge distributions exist for 2 expenses \((\d+) EUR and (\d+) EUR\)$"#)]
-async fn given_distributions_2_amounts(world: &mut FinancialWorld, _amount1: f64, amount2: f64) {
+// Les deux montants sont des ENTIERS d'euros dans le gabarit
+// (`\(\d+\) EUR`), et le second est lié tel quel à une colonne NUMERIC.
+// En `f64`, il traversait donc le flottant avant d'atteindre une colonne
+// exacte — ce que l'ADR-0008 §A proscrit, y compris dans un harnais.
+async fn given_distributions_2_amounts(world: &mut FinancialWorld, _amount1: u64, amount2: u64) {
     // Distribute first expense (already created in background)
     let uc = world
         .charge_distribution_use_cases
@@ -3758,7 +3762,7 @@ async fn given_distributions_2_amounts(world: &mut FinancialWorld, _amount1: f64
     .bind(id2)
     .bind(building_id)
     .bind(org_id)
-    .bind(amount2)
+    .bind(Decimal::from(amount2))
     .execute(pool)
     .await
     .expect("insert second expense");
